organize and refactor arm commands - move into subfolder - extend the abstract class CommandBase instead of implementing Command - replace the ArmSubsystem constructor argument with the ArmSubsystem#getInstance method
